Обновление внутренней архитектуры проекта * Глобальные переменные для оптимального доступа  вынесены в отдельный файл global_vars.py * Класс Button теперь импортируется из файла bytton.py
