+++ global_vars.py
@@ -1,0 +1,18 @@
+'''
+В данном файле содержатся глобальные переменные
+Для оптимального доступа из любого файла
+'''
+
+# Словарь с цветами, используемыми в коде
+COLORS = {
+    'main': '#222831',
+    'second_main': '#393E46',
+    'text': '#FFD369',
+    'second_text': '#EEEEEE',
+    'btn_inactive_color': "#393E46",
+    'btn_active_color': "#4f5257"
+}
+
+FONT = './Fonts/retro-land-mayhem.ttf'  # Шрифт
+ACTIVE_SCREEN = 'start'  # начальная сцена
+alreadyPressed = False  # Переменная для считывании состоянии кнопки
