+++ main.py
@@ -1,62 +1,10 @@
 import pygame
+from global_vars import *  # Импортируем глобальные переменные
+from button import Button  # Импортируем класс кнопки
 
 '''
 Objects
 '''
-
-
-class Button:
-    def __init__(self, width, height, inactive_color, active_color):
-        self.width = width  # Длина кнопки
-        self.height = height  # Ширина кнопки
-        self.inactive_color = inactive_color  # Базовый цвет
-        self.active_color = active_color  # Цвет при наведении
-
-    def draw(self, btn_x, btn_y, btn_text=None, action=None):
-        global alreadyPressed
-
-        mouse = pygame.mouse.get_pos()  # Позиция мыши
-        click = pygame.mouse.get_pressed()  # Обработка нажатия на кнопку мыши
-
-        # Если курсор в пределах кнопки..
-        if btn_x < mouse[0] < btn_x + self.width and \
-                btn_y < mouse[1] < btn_y + self.height:
-            # Отрисовываем кнопку с цветом <active_color>
-            pygame.draw.rect(
-                screen,
-                self.active_color,
-                (btn_x, btn_y, self.width, self.height)
-            )
-
-            # Если по ней кликнули ЛКМ и ключ action задан
-            if click[0] and action is not None:
-                alreadyPressed = True
-            else:
-                if alreadyPressed:
-                    action()
-                    alreadyPressed = False
-
-        # А ели за пределами кнопки..
-        else:
-            # Отрисовываем кнопку со стандартным цветом
-            pygame.draw.rect(
-                screen,
-                self.inactive_color,
-                (btn_x, btn_y, self.width, self.height)
-            )
-
-        _fontSize = 30
-        # Отрисовываем текст на кнопке
-        text_on_btn = pygame.font.Font(
-                                    FONT,
-                                    _fontSize
-                                    )\
-            .render(btn_text, True, COLORS['second_text'])
-        screen.blit(
-            text_on_btn,
-            (btn_x + (self.width // 2 - text_on_btn.get_width() // 2),
-                btn_y + (self.height // 2 - text_on_btn.get_height() // 2))
-                )
 
 
 def start_scr_loader():
@@ -86,7 +34,7 @@ def start_screen(screen, text_on_screen, text_on_button):
 
     # Создаем экземпляр класса Button (кнопку на экране завершения игры)
     btn = Button(
-        btn_width, btn_height,
+        screen, btn_width, btn_height,
         COLORS['btn_inactive_color'],
         COLORS['btn_active_color']
         )
@@ -124,7 +72,7 @@ def lvl_1(screen):
 
     # Создаем экземпляр класса Button (кнопку на экране завершения игры)
     btn = Button(
-        btn_width, btn_height,
+        screen, btn_width, btn_height,
         COLORS['btn_inactive_color'],
         COLORS['btn_active_color']
         )
@@ -162,7 +110,7 @@ def lvl_2(screen):
 
     # Создаем экземпляр класса Button (кнопку на экране завершения игры)
     btn = Button(
-        btn_width, btn_height,
+        screen, btn_width, btn_height,
         COLORS['btn_inactive_color'],
         COLORS['btn_active_color']
         )
@@ -174,7 +122,7 @@ def lvl_2(screen):
         "Go Win ->", win_scr_loader)
 
     btn2 = Button(
-        btn_width, btn_height,
+        screen, btn_width, btn_height,
         COLORS['btn_inactive_color'],
         COLORS['btn_active_color']
         )
@@ -218,7 +166,7 @@ def end_screen(screen, text_on_screen, text_on_button):
 
     # Создаем экземпляр класса Button (кнопку на экране завершения игры)
     btn = Button(
-        btn_width, btn_height,
+        screen, btn_width, btn_height,
         COLORS['btn_inactive_color'],
         COLORS['btn_active_color']
         )
@@ -228,25 +176,6 @@ def end_screen(screen, text_on_screen, text_on_button):
         width // 2 - btn_width // 2,
         height // 2 - btn_height // 2 + 75,
         text_on_button, start_scr_loader)
-
-
-'''
-Variables
-'''
-
-# Словарь с цветами, используемыми в коде
-COLORS = {
-    'main': '#222831',
-    'second_main': '#393E46',
-    'text': '#FFD369',
-    'second_text': '#EEEEEE',
-    'btn_inactive_color': "#393E46",
-    'btn_active_color': "#4f5257"
-}
-
-FONT = './Fonts/retro-land-mayhem.ttf'  # Шрифт
-ACTIVE_SCREEN = 'start'  # начальная сцена
-alreadyPressed = False  # Переменная для считывании состоянии кнопки
 
 
 """
